+++ MasterCube/src/Settings.java
@@ -1,5 +1,5 @@
 public class Settings{
-	public int[] options = new int[3];
+	public byte[] options = new byte[3];
 	public boolean exit = false;
 	public final String gameOptions[] = { "new game", "save game", "load game" };
 	public final String fieldsOptions[] = { "4 fields", "6 fields" };
